+++ src/parser/inline/extension/index.test.ts
@@ -24,9 +24,10 @@ describe('Unit: parser/inline/extension/index', () => {
     it('basic', () => {
       assert.deepStrictEqual(inspect(parser('[#a]')), [['<a href="#index:a" rel="noopener" class="index">a</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[#a ]')), [['<a href="#index:a" rel="noopener" class="index">a</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[#a b]')), [['<a href="#index:a-b" rel="noopener" class="index">a b</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#a b]')), [['<a href="#index:a_b" rel="noopener" class="index">a b</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#a  b]')), [['<a href="#index:a_b" rel="noopener" class="index">a  b</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[#a\\ ]')), [['<a href="#index:a" rel="noopener" class="index">a</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('[#a\\ b]')), [['<a href="#index:a-b" rel="noopener" class="index">a b</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#a\\ b]')), [['<a href="#index:a_b" rel="noopener" class="index">a b</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[#[]]')), [['<a href="#index:[]" rel="noopener" class="index">[]</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[#\\]]')), [['<a href="#index:]" rel="noopener" class="index">]</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[#\\\\]')), [['<a href="#index:\\" rel="noopener" class="index">\\</a>'], '']);