+++ src/parser/inline/extension/indexer.test.ts
@@ -20,7 +20,8 @@ describe('Unit: parser/inline/extension/indexer', () => {
     it('valid', () => {
       assert.deepStrictEqual(inspect(parser(' [#a]')), [['<span class="indexer" data-index="a"></span>'], '']);
       assert.deepStrictEqual(inspect(parser(' [#a] ')), [['<span class="indexer" data-index="a"></span>'], '']);
-      assert.deepStrictEqual(inspect(parser(' [#a b]')), [['<span class="indexer" data-index="a-b"></span>'], '']);
+      assert.deepStrictEqual(inspect(parser(' [#a b]')), [['<span class="indexer" data-index="a_b"></span>'], '']);
+      assert.deepStrictEqual(inspect(parser(' [#a  b]')), [['<span class="indexer" data-index="a_b"></span>'], '']);
       assert.deepStrictEqual(inspect(parser(' [#A]')), [['<span class="indexer" data-index="A"></span>'], '']);
       assert.deepStrictEqual(inspect(parser(' [#*A*]')), [['<span class="indexer" data-index="A"></span>'], '']);
       assert.deepStrictEqual(inspect(parser(' [#`A`]')), [['<span class="indexer" data-index="`A`"></span>'], '']);